Fix hypervisor native: TLS vcpu leak, thread safety, and code cleanup
- Fix vcpu resource leak in nativeDestroy by manually cleaning up TLS before pthread_key_delete
- Replace volatile bool with std::atomic<bool> for stop_request thread safety
- Simplify debug register switch-case (~230 lines) into lookup table arrays
- Make HYP_ASSERT_SUCCESS independent of NDEBUG so it works in Release builds
- Cache find_vcpus() result to avoid repeated Mach-O symbol table traversal
- Fix syn_get_ec parameter type from uint32_t to uint64_t to match caller

Made-with: Cursor

+++ backend/hypervisor/src/main/native/hypervisor/hypervisor.h
@@ -63,8 +63,8 @@ enum arm_exception_class {
 #define ARM_EL_IL (1 << ARM_EL_IL_SHIFT)
 #define ARM_EL_ISV (1 << ARM_EL_ISV_SHIFT)
 
-static uint32_t syn_get_ec(uint32_t syn) {
-  return syn >> ARM_EL_EC_SHIFT;
+static uint32_t syn_get_ec(uint64_t syn) {
+  return (uint32_t) (syn >> ARM_EL_EC_SHIFT);
 }
 
 #define PSR_MODE_EL0t	0x00000000
